refactor: dedupe luminance + action-dispatch logic (CodeRabbit)
- Extract the renderer's canvas-normalize + lightness test into a single
  `isColorLight` helper (`lib/colorScheme.ts`), reused by `useTitleBarIdentity`
  and `TitlePopupApp` instead of two copies; it reuses the shared
  `perceivedLuminance` math
- Collapse the guarded `emit('trigger-action', …)` repeated across the update +
  migrate pills into one `triggerInstallAction` method in `ChooserInstallTile`
