+++ src/renderer/src/comfyTitleBar/useTitleBarIdentity.ts
@@ -7,7 +7,7 @@ import {
   isLoadingLockdownMode,
   type FirstUseMode,
 } from '../../../shared/firstUseMode'
-import { perceivedLuminance, LUMINANCE_LIGHT_THRESHOLD } from '../../../shared/colorLuminance'
+import { isColorLight } from '../lib/colorScheme'
 
 interface InstallTypeMeta {
   icon: ReturnType<typeof installTypeMetaFor>['icon']
@@ -90,22 +90,8 @@ export function useTitleBarIdentity(opts: UseTitleBarIdentityOpts): TitleBarIden
   const showBrandMark = computed(() => opts.isInstallLess.value && !isPreviewMode.value)
 
   /** True when the reported ComfyUI bg is light, so the title bar's `.is-light` chrome
-   *  variants (lighter hover/pills/chips) kick in to stay legible on the matching surface.
-   *  Normalises any CSS color to hex via the canvas, then runs the standard perceived-luminance
-   *  test; defaults to dark on a non-hex / missing color. */
-  const isLight = computed(() => {
-    const bg = themeBg.value
-    if (!bg) return false
-    const ctx = document.createElement('canvas').getContext('2d')
-    if (!ctx) return false
-    ctx.fillStyle = bg
-    const hex = ctx.fillStyle as string
-    if (!hex.startsWith('#') || hex.length < 7) return false
-    const r = parseInt(hex.slice(1, 3), 16)
-    const g = parseInt(hex.slice(3, 5), 16)
-    const b = parseInt(hex.slice(5, 7), 16)
-    return perceivedLuminance(r, g, b) >= LUMINANCE_LIGHT_THRESHOLD
-  })
+   *  variants (lighter hover/pills/chips) kick in to stay legible on the matching surface. */
+  const isLight = computed(() => isColorLight(themeBg.value))
 
   let unsubTitle: (() => void) | undefined
   let unsubSourceCategory: (() => void) | undefined
