+++ src/renderer/src/comfyTitlePopup/TitlePopupApp.vue
@@ -11,7 +11,7 @@ import { dismissPickerModals } from './dismissPickerModals'
 import { popupLocaleSource } from './pickerSettingsApiShim'
 import { useAppLocale } from '../lib/useAppLocale'
 import type { DetailSection, SnapshotListData } from '../types/ipc'
-import { perceivedLuminance, LUMINANCE_LIGHT_THRESHOLD } from '../../../shared/colorLuminance'
+import { isColorLight } from '../lib/colorScheme'
 
 // Title-bar dropdown popup shell. Hosts every title-bar dropdown in one
 // reused transparent WebContentsView attached to the host window. Each open
@@ -222,17 +222,7 @@ const globalSettingsSnapshot = ref<GlobalSettingsSnapshot>({
 const downloadsState = ref<DownloadsState>({ active: [], recent: [] })
 
 /** Body-luminance test driving is-light styling; matches TitleBarApp.vue. */
-const isLight = computed(() => {
-  const ctx = document.createElement('canvas').getContext('2d')
-  if (!ctx) return false
-  ctx.fillStyle = themeBg.value
-  const hex = ctx.fillStyle as string
-  if (!hex.startsWith('#') || hex.length < 7) return false
-  const r = parseInt(hex.slice(1, 3), 16)
-  const g = parseInt(hex.slice(3, 5), 16)
-  const b = parseInt(hex.slice(5, 7), 16)
-  return perceivedLuminance(r, g, b) >= LUMINANCE_LIGHT_THRESHOLD
-})
+const isLight = computed(() => isColorLight(themeBg.value))
 
 function handleActivate(id: string): void {
   bridge?.activate(id)
